# ProFind/Lib/Global/Services/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebService",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "operationId": "GetClient",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "operationId": "PutClient",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "operationId": "DeleteClient",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Professionals/{id}": {
      "get": {
        "tags": [
          "Professionals"
        ],
        "operationId": "GetProfessional",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Professional"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Professionals"
        ],
        "operationId": "PutProfessional",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Professional"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Professionals"
        ],
        "operationId": "DeleteProfessional",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Professions/{id}": {
      "get": {
        "tags": [
          "Professions"
        ],
        "operationId": "GetProfession",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Profession"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Professions"
        ],
        "operationId": "PutProfession",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Profession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Professions"
        ],
        "operationId": "DeleteProfession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Securityanswerclients": {
      "get": {
        "tags": [
          "Securityanswerclients"
        ],
        "operationId": "GetSecurityanswerclients",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Securityanswerclient"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Securityanswerclients"
        ],
        "operationId": "PostSecurityanswerclient",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Securityanswerclient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Securityanswerclient"
            }
          }
        }
      }
    },
    "/api/Securityanswerclients/{id}": {
      "get": {
        "tags": [
          "Securityanswerclients"
        ],
        "operationId": "GetSecurityanswerclient",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Securityanswerclient"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Securityanswerclients"
        ],
        "operationId": "PutSecurityanswerclient",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Securityanswerclient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Securityanswerclients"
        ],
        "operationId": "DeleteSecurityanswerclient",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Securityanswerprofessionals": {
      "get": {
        "tags": [
          "Securityanswerprofessionals"
        ],
        "operationId": "GetSecurityanswerprofessionals",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Securityanswerprofessional"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Securityanswerprofessionals"
        ],
        "operationId": "PostSecurityanswerprofessional",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Securityanswerprofessional"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Securityanswerprofessional"
            }
          }
        }
      }
    },
    "/api/Securityanswerprofessionals/{id}": {
      "get": {
        "tags": [
          "Securityanswerprofessionals"
        ],
        "operationId": "GetSecurityanswerprofessional",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Securityanswerprofessional"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Securityanswerprofessionals"
        ],
        "operationId": "PutSecurityanswerprofessional",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Securityanswerprofessional"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Securityanswerprofessionals"
        ],
        "operationId": "DeleteSecurityanswerprofessional",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Activitycomment": {
      "type": "object",
      "properties": {
        "idAc": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "commentAc": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "dateAc": {
          "$ref": "#/definitions/DateOnly"
        },
        "askToCheckChatAc": {
          "type": "boolean"
        },
        "idA1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idA1Navigation": {
          "$ref": "#/definitions/Activity"
        },
        "idC5Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP5Navigation": {
          "$ref": "#/definitions/Professional"
        }
      }
    },
...
    "Changepasswordcode": {
      "type": "object",
      "properties": {
        "idCpc": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "codeCpc": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "verifiedCpc": {
          "type": "boolean"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        }
      }
    },
    "Client": {
      "type": "object",
      "properties": {
        "idC": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameC": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "emailC": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "passwordC": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "pictureC": {
          "format": "byte",
          "type": "string"
        },
        "activitycomments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activitycomment"
          }
        },
        "changepasswordcodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Changepasswordcode"
          }
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "proposalnotifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposalnotification"
          }
        },
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposal"
          }
        },
        "securityanswerclients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Securityanswerclient"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "idDp": {
          "format": "int32",
          "type": "integer"
        },
        "nameDp": {
          "maxLength": 30,
          "minLength": 0,
          "type": "string"
        },
        "professionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Professional"
          }
        }
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "idM": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "contentM": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "imageM": {
          "format": "byte",
          "type": "string"
        },
        "documentM": {
          "format": "byte",
          "type": "string"
        },
        "locationM": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "audioM": {
          "format": "byte",
          "type": "string"
        },
        "dateTimeM": {
          "$ref": "#/definitions/DateOnly"
        },
        "idP4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP4Navigation": {
          "$ref": "#/definitions/Professional"
        }
      }
    },
...
    "Profession": {
      "type": "object",
      "properties": {
        "idPfs": {
          "format": "int32",
          "type": "integer"
        },
        "namePfs": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "professionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Professional"
          }
        }
      }
    },
    "Professional": {
      "type": "object",
      "properties": {
        "idP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameP": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "dateBirthP": {
          "$ref": "#/definitions/DateOnly"
        },
        "emailP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "passwordP": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "activeP": {
          "type": "boolean"
        },
        "sexP": {
          "type": "boolean"
        },
        "duip": {
          "maxLength": 15,
          "minLength": 0,
          "type": "string"
        },
        "afpp": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "isssp": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "zipCodeP": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "salaryP": {
          "format": "float",
          "type": "number"
        },
        "hiringDateP": {
          "$ref": "#/definitions/DateOnly"
        },
        "pictureP": {
          "format": "byte",
          "type": "string"
        },
        "curriculumP": {
          "format": "byte",
          "type": "string"
        },
        "latitudeLocationP": {
          "format": "double",
          "type": "number"
        },
        "longitudeLocationP": {
          "format": "double",
          "type": "number"
        },
        "idPfs1": {
          "format": "int32",
          "type": "integer"
        },
        "idDp1": {
          "format": "int32",
          "type": "integer"
        },
        "idWdt1": {
          "format": "int32",
          "type": "integer"
        },
        "idDp1Navigation": {
          "$ref": "#/definitions/Department"
        },
        "idPfs1Navigation": {
          "$ref": "#/definitions/Profession"
        },
        "activitycomments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activitycomment"
          }
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "projecttemplates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projecttemplate"
          }
        },
        "proposalnotifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposalnotification"
          }
        },
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposal"
          }
        },
        "securityanswerprofessionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Securityanswerprofessional"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "Project": {
      "type": "object",
      "properties": {
        "idPj": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titlePj": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPj": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "picturePj": {
          "format": "byte",
          "type": "string"
        },
        "totalPricePj": {
          "format": "float",
          "type": "number"
        },
        "idPs1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP1Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPs1Navigation": {
          "$ref": "#/definitions/Projectstatus"
        },
        "activities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activity"
          }
        },
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Notification"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          }
        }
      }
    },
    "Projectpay": {
      "type": "object",
      "properties": {
        "idPpy": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "percentagePpy": {
          "format": "float",
          "type": "number"
        },
        "conceptPpy": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "amountPpy": {
          "format": "float",
          "type": "number"
        },
        "currencyPpy": {
          "maxLength": 3,
          "minLength": 0,
          "type": "string"
        },
        "hasLimitDatePpy": {
          "type": "boolean"
        },
        "limitDatePpy": {
          "$ref": "#/definitions/DateOnly"
        },
        "defaultAmountPpy": {
          "format": "float",
          "type": "number"
        },
        "payStatusPpy": {
          "type": "string"
        },
        "idP3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPj3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP3Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPj3Navigation": {
          "$ref": "#/definitions/Project"
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
...
    "Proposalnotification": {
      "type": "object",
      "properties": {
        "idPn": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "contentPn": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "imagePn": {
          "format": "byte",
          "type": "string"
        },
        "actionPn": {
          "type": "string"
        },
        "idPp1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP4Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPp1Navigation": {
          "$ref": "#/definitions/Proposal"
        }
      }
    },
...
    "Securityanswerclient": {
      "type": "object",
      "properties": {
        "idSa": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "answerSa": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "idSq1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idSq1Navigation": {
          "$ref": "#/definitions/Securityquestion"
        }
      }
    },
    "Securityanswerprofessional": {
      "type": "object",
      "properties": {
        "idSa": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "answerSa": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "idSq": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPNavigation": {
          "$ref": "#/definitions/Professional"
        },
        "idSqNavigation": {
          "$ref": "#/definitions/Securityquestion"
        }
      }
    },
...
  }
}